requirements.txt
apache-airflow==2.9.3
pandas==2.2.3
plotly==6.0.1
streamlit==1.44.1

